[kernel] Panic quickly for unexpected page faults in kernel mode
The only kind of page fault that should ever happen in kernel mode
is for a data access during explicit user-copy operations.  For any
other case, a quick panic with full details of the fault site is
crucial to reducing the potential cascade harm of the bug and to
identifying the underlying root cause of the fault.

Bug: 66141
Change-Id: I4e4dce26f3e14733d7be7c07012aafca586f53c5