[kernel][vm] Sync executable pages up front
This calls out explicitly the requirement that users not modify VMOs
after calling zx_vmo_replace_as_executable, unless they synchronize the
instruction cache explicitly. The requirement was already implicit as
once mappings are created the kernel has no way to keep the executable
pages in sync. Additionally it is made explicit that only cached VMOs
may be made executable.

Performing the cache sync operation on the page fault path is both
racy, as currently implemented, and overly pessimistic. The race exists
due to another thread from the same address space able to use the mmu
mapping prior to the cache sync from being performed. The pessimism
stems from the data cache clean happening every time a mapping occurs,
when it only needs to happen once after a given page is modified.

Although other solutions exist to resolve the race, these require
creating additional read-only mappings in the map or unmap path and
hence additional break-before-make steps.

The solution here is to eagerly sync any pages present in a VMO at the
point in time it becomes executable, alleviating the need to do it
when any particular mapping faults in the page. For user pager backed
VMOs any content faulted in via SupplyPages is notionally part of the
original content, and it is the fault of the kernel for evicting the
backing page, and so in this case the kernel will perform sync and
the user pager is not required to. To achieve this the VM objects
remember if they have been marked executable and sync any supplied
pages if they have.

By performing the sync in these locations the only option to invalidate
the instruction cache when it is VIPT is to invalidate the entire
instruction cache. As creating executable VMOs, and supplying pages to
executable VMOs, is not as common as faulting in mappings this is not
necessarily a performance regressions.

Testing does not show any noticable macro performance implications
over the existing incorrect racy implementation. Memory bandwidth usage
in particular does not seem to be increased despite the increased cache
invalidations.

Change-Id: I4e8ab6e9facd73ab7b1e2bdf3b86b7c3ed41cb77