[kernel][arm64][mmu] use 4K page granules for the bootstrap trampoline
Currently the start.S code is more clever than it probably should be and
compile time decides wehat size page granules to use to map the large
(1GB/512MB) page around the current code at boot to trampoline to the
high address. It's currently set to use 64KB page granules in the
temporary address space it uses. This is generally fine except 64K pages
are technically optional according to the ARM ARM.

Switch to using 4K pages always which are more or less mandatory, or at
least mandatory to zircon. This avoids any issues booting on cpus that
don't support 64K pages.

This reduces the size of the physical address space the kernel needs to
be physically located in at boot time, but 39 bits is still pretty good,
and it's unlikely we'll encounter a cpu where the kernel is loaded
outside of that range.

Change-Id: Idc962ebf39d689ec124f20392f910d38824c0b3c