[wlan] Retry sending packets until success in sim-client-vs-sim-ap
There is no need for sim-client-vs-sim-ap to test if either side receives
the first packet sent since the sender, even in reality, cannot know if
the receiver is ready yet to receive packets immediately after ESSSA is
established. This is because the controlled port may not be open yet on
the other end.

To account for the controlled port potentially being closed before the
first packet is sent, transmission of packets should be retried until
acknowledged or test timeout.

Bug: 72109
Fixed: 73544
Test: fx test sim-client-vs-sim-ap
Multiply: sim-client-vs-sim-ap

Change-Id: I1915055f58d5a4dcab176cc5bf149b2ac957ec90