[wlan] Synchronize sim-client-vs-sim-ap channel behavior
The client and ap both use softmac to communicate and therefore the
sim-client-vs-sim-ap test can leverage almost all of the softmac
implemenatation for communcation between the client and ap. The only
catch in the test is that the ap and client should only relay packets
between each other when on the same channel, as they would in a
real-world environment.

Test: fx test sim-client-vs-sim-ap
Bug: 73547
Change-Id: I85b7809507633f0d734b2d72544ca7265224ae81